api-gateway/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
python-multipart
